requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
chardet==4.0.0
idna==3.2
MouseInfo==0.1.3
multidict==5.1.0
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.27
python-engineio==4.2.1
python-socketio==5.4.0
python3-xlib==0.15
PyTweening==1.0.3
typing-extensions==3.10.0.0
yarl==1.6.3
